refactor fuzz test
